refactor: extract job schema registration helpers
